Añadida refactorizacion de codigo y nuevos comentarios

+++ src/main/java/com/david/edulib/service/implementation/AutorServiceImpl.java
@@ -2,7 +2,6 @@ package com.david.edulib.service.implementation;
 
 import java.util.List;
 import java.util.Optional;
-
 import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Service;
 import com.david.edulib.dto.AutorDTO;
@@ -34,6 +33,9 @@ public class AutorServiceImpl implements AutorService{
     public Optional<AutorDTO> obtenerPorId(Integer id) {
         // OJO, arriba el map es el de stream, porque tenemos una lista, aqui es un map de optional, que trabaja sobre nuestra clase pero 
         // devuelve un optional igualmente
+        if (id == null){
+            return Optional.empty();
+        }
         return autorRepository.findById(id).map(autorMapper::toDTO);
     }
 
@@ -45,16 +47,20 @@ public class AutorServiceImpl implements AutorService{
 
     @Override
     @Transactional // Poruqe escribe en la base de datos (no es un select)
-    public AutorDTO guardar(AutorDTO autorDTO) throws AutorException{
+    public AutorDTO guardar(AutorDTO autorDTO){
+        if (autorDTO == null){
+            throw new AutorException("Intentado registrar un autor sin datos");
+        }
         Autor autor;
-        // Primero, si viene con ID, vamos a proceder con una actualizacion de los datos del repositorio
+        // Edicion, viene con Id
         if (autorDTO.getId()!=null){
-            // Si no existe lanzamos exception, sino continuamos con la logica de edicion
+            // Si no existe lanzamos exception, sino continuamos con la logica de edicion, la lista de libros me la trae la relacion EAGER
             autor = autorRepository.findByNombre(autorDTO.getNombre()).orElseThrow(() -> new AutorException("No existe el autor con id "+autorDTO.getId()));
             autor.setNombre(autorDTO.getNombre());
             autor.setNacionalidad(autorDTO.getNacionalidad());
+        // Creación, viene sin Id    
         } else {
-            // Si no, es que el autor no trae ID, con lo cual es una nueva entrada
+            // El arraylist de libros se inicializa solo con @Builder.Default de autor
             autor = autorRepository.save(autorMapper.toEntity(autorDTO));
         }
         return autorMapper.toDTO(autor);
@@ -63,18 +69,24 @@ public class AutorServiceImpl implements AutorService{
     @Override
     @Transactional
     public void borrarPorId(Integer id){
-        autorRepository.deleteById(id);
+        if (id != null){    
+            autorRepository.deleteById(id);
+        }
     }
 
     @Override 
     @Transactional
     public void borrar(AutorDTO autorDTO){
-        autorRepository.delete(autorMapper.toEntity(autorDTO));
+        if (autorDTO != null){
+            autorRepository.delete(autorMapper.toEntity(autorDTO));
+        }
     }
 
     @Override
     @Transactional
     public void guardar(Autor autor){
-        autorRepository.save(autor);
+        if (autor != null){
+            autorRepository.save(autor);
+        }
     }
 }
